feat: enhance user authentication and authorization checks across handlers
- Added error handling for user ID extraction from token in multiple handlers to ensure unauthorized access is properly managed.
- Updated DeleteGymRoutine function to verify gym admin status before allowing deletion.
- Improved error messages for better clarity in case of authentication failures.
- Refactored user ID retrieval to return errors, enhancing overall robustness of the authentication process.

+++ internal/database/gyms-routines.go
@@ -1,6 +1,8 @@
 package database
 
 import (
+	"errors"
+
 	gonanoid "github.com/matoous/go-nanoid/v2"
 	"github.com/xyztavo/go-gym/internal/models"
 )
@@ -33,7 +35,18 @@ func GetGymRoutines(gymId string) (gymRoutines []models.Routine, err error) {
 	return gymRoutines, nil
 }
 
-func DeleteGymRoutine(id string) (err error) {
+func DeleteGymRoutine(adminId string, id string) (err error) {
+	admin, err := GetUserById(adminId)
+	if err != nil {
+		return err
+	}
+	gym, err := GetGymById(*admin.GymId)
+	if err != nil {
+		return err
+	}
+	if gym.AdminId != adminId {
+		return errors.New("user is not the admin of the gym")
+	}
 	_, err = db.Exec("DELETE FROM gyms_routines WHERE id = $1", id)
 	return err
 }
